+++ llama_index/llms/localai.py
@@ -56,14 +56,14 @@ class LocalAI(OpenAI):
         # This subclass only supports max_tokens via LocalAI(..., max_tokens=123)
         if self.max_tokens is not None:
             return
-        all_kwargs.pop("max_tokens", None)
 
     @property
     def _is_chat_model(self) -> bool:
         if self.globally_use_chat_completions is not None:
             return self.globally_use_chat_completions
         raise NotImplementedError(
-            f"Inferring of /chat/completions is not supported by {type(self).__name__}."
-            f" Please use the kwarg 'use_chat_completions' in your query, setting"
-            f" True to use /chat/completions or False to use /completions."
+            "Inferring of when to use /chat/completions is unsupported by"
+            f" {type(self).__name__}. Please either set 'globally_use_chat_completions'"
+            " arg during construction, or pass the arg 'use_chat_completions' in your"
+            " query, setting True for /chat/completions or False for /completions."
         )